initial commit, minor changes
